n2mobile/requirement.txt
asgiref==3.4.1
Django==3.2.9
django-filter==21.1
django-passwords==0.3.12
djangorestframework==3.12.4
djangorestframework-gis==0.17
postgis==1.0.4
psycopg2==2.9.1
psycopg2-binary==2.9.1
python-dotenv==0.19.1
pytz==2021.3
sqlparse==0.4.2
